+++ src/shader/mod.rs
@@ -1,11 +1,9 @@
 pub mod constants;
-pub mod glsl;
 pub mod program;
 pub mod read;
 pub mod service;
 
 pub use constants::*;
-pub use glsl::*;
 pub use program::*;
 pub use read::*;
 pub use service::*;
