+++ src/shader/program.rs
@@ -1,4 +1,20 @@
-use crate::{cstr_with_len, Shader};
+use std::ffi::CString;
+
+pub fn cstr_with_len(len: usize) -> CString {
+    let mut buffer: Vec<u8> = Vec::with_capacity(len + 1);
+    buffer.extend([b' '].iter().cycle().take(len));
+    unsafe { CString::from_vec_unchecked(buffer) }
+}
+
+pub fn cstr_to_str(cstr: &CString) -> String {
+    cstr.to_string_lossy().to_string()
+}
+
+#[derive(Debug)]
+pub enum ShaderError {
+    CompilationError { error: String },
+    FileError { error: String },
+}
 
 /// TODO: Should only be the data structure and not actually construct the glsl shader
 /// TODO: Merge with struct Shader. No reason to keep two of them around
@@ -8,7 +24,7 @@ pub struct ShaderProgram {
 }
 
 impl ShaderProgram {
-    pub fn new(vert_shader: Shader, frag_shader: Shader) -> Self {
+    pub fn new(vert_shader: ShaderProgram, frag_shader: ShaderProgram) -> Self {
         let id = unsafe { gl::CreateProgram() };
         unsafe {
             gl::AttachShader(id, vert_shader.id);
@@ -48,6 +64,14 @@ impl ShaderProgram {
 
         Self { id }
     }
+
+    pub fn from_source(
+        source: String,
+        shader_type: gl::types::GLuint,
+    ) -> anyhow::Result<Self, ShaderError> {
+        let id = shader_from_string(source, shader_type)?;
+        Ok(ShaderProgram { id })
+    }
 }
 
 impl Drop for ShaderProgram {
@@ -56,4 +80,53 @@ impl Drop for ShaderProgram {
             gl::DeleteProgram(self.id);
         }
     }
+}
+
+fn shader_from_string(
+    source: String,
+    shader_type: gl::types::GLuint,
+) -> anyhow::Result<gl::types::GLuint, ShaderError> {
+    let c_src = CString::new(source).expect("Could not convert source to CString");
+
+    // check for includes
+
+    let id = unsafe { gl::CreateShader(shader_type) };
+
+    unsafe {
+        gl::ShaderSource(id, 1, &c_src.as_ptr(), std::ptr::null());
+        gl::CompileShader(id);
+    }
+
+    let mut success: gl::types::GLint = 1;
+    unsafe {
+        gl::GetShaderiv(id, gl::COMPILE_STATUS, &mut success);
+    }
+
+    // in case something failed we want to extract the error and return it
+    if success == 0 {
+        // fetch the required buffer length
+        let mut len: gl::types::GLint = 0;
+        unsafe {
+            gl::GetShaderiv(id, gl::INFO_LOG_LENGTH, &mut len);
+        }
+
+        let error = cstr_with_len(len as usize);
+
+        unsafe {
+            gl::GetShaderInfoLog(
+                id,
+                len,
+                std::ptr::null_mut(),
+                error.as_ptr() as *mut gl::types::GLchar,
+            )
+        }
+
+        // Prints the compilation error to console
+        log::error!("{}", error.to_string_lossy());
+        return Err(ShaderError::CompilationError {
+            error: cstr_to_str(&error),
+        });
+    }
+
+    Ok(id)
 }
