+++ src/utils.rs
@@ -1,15 +1,3 @@
-use std::ffi::CString;
-
-pub fn cstr_with_len(len: usize) -> CString {
-    let mut buffer: Vec<u8> = Vec::with_capacity(len + 1);
-    buffer.extend([b' '].iter().cycle().take(len));
-    unsafe { CString::from_vec_unchecked(buffer) }
-}
-
-pub fn cstr_to_str(cstr: &CString) -> String {
-    cstr.to_string_lossy().to_string()
-}
-
 pub fn string_between<'v>(value: &'v str, start: &str, end: &str) -> &'v str {
     if let Some(start_idx) = value.rfind(start) {
         if let Some(end_idx) = value.rfind(end) {
